refactor: rename test files and adjust package to assessmentplan in integration tests
