+++ internal/api/handler/oscal/assessmentplan/task_integration_test.go
@@ -1,6 +1,6 @@
 //go:build integration
 
-package test
+package assessmentplan
 
 import (
 	"bytes"
